lastly updated by Siva

+++ InventryManagementSystem/src/main/java/com/project/IMS/repository/CategoryRepository.java
@@ -1,12 +1,9 @@
 package com.project.IMS.repository;
 
+import com.project.IMS.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.project.IMS.entity.Category;
-
 @Repository
-// Siva's Module
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-	
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
